requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
fastapi==0.135.3
flatbuffers==25.12.19
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.0
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
pymupdf-layout==1.27.2.2
pymupdf4llm==1.27.2.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
starlette==1.0.0
sympy==1.14.0
tabulate==0.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
